Refactor For Simpler Imports
This is not a production library but something as practice, so rather than long complex imports I want to do it simply. So org name was removed for simpler code.
